Fix image url to string (#31)
## Goal of the PR

The goal of the PR is to make ImageUrl.rawUrl visible again. It may be
usefull for integrators that use Room or other string based data
storage.

The rawUrl is was never really hidden has an integrator can create an
`ImageUrlDecorator` that return the rawUrl.

The decision is to revert previous modifications.