Fix Maven publication (#64)
There was an issue with Maven publication when using Java 17.
I wasn't able to figure out the exact cause, but using Java 11 seems to
work fine, so I've updated the project to use that version.
I've also updated the `code_quality.yml` workflow to add a step to
publish to Maven local, so we can always ensure that publication works.

The error that was happening can be seen in the build log of the first
commit:
https://github.com/SRGSSR/srgdataprovider-android/actions/runs/12274751507/job/34248321537?pr=64